Moving Mandelbrot and Julia calculatin procedures in separate modules
