+++ GTFractals/GTFractals.vcxproj
@@ -238,11 +238,15 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="DialogProc.cpp" />
+    <ClCompile Include="Julia.cpp" />
     <ClCompile Include="main.cpp" />
+    <ClCompile Include="Mandelbrot.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="dialog.h" />
     <ClInclude Include="global.h" />
+    <ClInclude Include="Julia.h" />
+    <ClInclude Include="Mandelbrot.h" />
     <ClInclude Include="resource.h" />
   </ItemGroup>
   <ItemGroup>
