+++ GTFractals/GTFractals.vcxproj.filters
@@ -21,6 +21,12 @@
     <ClCompile Include="DialogProc.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="Mandelbrot.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="Julia.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="resource.h">
@@ -30,6 +36,12 @@
       <Filter>Resource Files</Filter>
     </ClInclude>
     <ClInclude Include="global.h">
+      <Filter>Resource Files</Filter>
+    </ClInclude>
+    <ClInclude Include="Mandelbrot.h">
+      <Filter>Resource Files</Filter>
+    </ClInclude>
+    <ClInclude Include="Julia.h">
       <Filter>Resource Files</Filter>
     </ClInclude>
   </ItemGroup>
