+++ GTFractals/Julia.h
@@ -1,0 +1,10 @@
+#ifndef JULIA_H
+#define JULIA_H
+
+#include <windows.h>
+
+#include "global.h"
+
+void DrawJulia(HWND); // Draw the Julia set
+
+#endif
