+++ GTFractals/Mandelbrot.cpp
@@ -1,0 +1,36 @@
+#include "Mandelbrot.h"
+
+
+// Function to draw the Mandelbrot set
+void DrawMandelbrot(HWND hwnd) {
+	PAINTSTRUCT ps;
+	HDC hdc = BeginPaint(hwnd, &ps);
+
+	RECT rect;
+	GetClientRect(hwnd, &rect);
+	int width = rect.right;
+	int height = rect.bottom;
+
+	for (int py = 0; py < height; ++py) {
+		double y0 = ymin + (ymax - ymin) * py / height;
+		for (int px = 0; px < width; ++px) {
+			double x0 = xmin + (xmax - xmin) * px / width;
+
+			double x = 0.0;
+			double y = 0.0;
+			int iteration = 0;
+
+			while (x * x + y * y <= 4.0 && iteration < max_iter) {
+				double xtemp = x * x - y * y + x0;
+				y = 2 * x * y + y0;
+				x = xtemp;
+				++iteration;
+			}
+
+			COLORREF color = (iteration == max_iter) ? RGB(0, 0, 0) : RGB(255 - (iteration * 10) % 255, 255 - (iteration * 5) % 255, 255);
+			SetPixel(hdc, px, py, color);
+		}
+	}
+
+	EndPaint(hwnd, &ps);
+}
