+++ GTFractals/Mandelbrot.h
@@ -1,0 +1,10 @@
+#ifndef MANDELBROT_H
+#define MANDELBROT_H
+
+#include <windows.h>
+
+#include "global.h"
+
+void DrawMandelbrot(HWND); // Draw the Mandelbrot set
+
+#endif
