+++ GTFractals/main.cpp
@@ -5,6 +5,9 @@
 #include "resource.h"
 #include "global.h"
 
+#include "Mandelbrot.h"
+#include "Julia.h"
+
 
 // Global variables
 double xmin = -2.25, ymin = -1.5, xmax = 0.75, ymax = 1.5;	// Initial coordinates
@@ -26,8 +29,6 @@ typedef struct {
 
 // Function prototypes for the main window
 void AddMenus(HWND);			// Add menus to the window
-void DrawMandelbrot(HWND);		// Draw the Mandelbrot set
-void DrawJulia(HWND);			// Draw the Julia set
 void UpdateFractal(HWND);		// Update the fractal based on the selection rectangle
 void DrawSelectionRect(HWND);	// Draw the selection rectangle
 
@@ -122,76 +123,6 @@ void AddMenus(HWND hwnd) {
 	AppendMenu(hMenu, MF_POPUP, (UINT_PTR)hHelpMenu, L"?");
 
 	SetMenu(hwnd, hMenu);
-}
-
-// Function to draw the Mandelbrot set
-void DrawMandelbrot(HWND hwnd) {
-	PAINTSTRUCT ps;
-	HDC hdc = BeginPaint(hwnd, &ps);
-
-	RECT rect;
-	GetClientRect(hwnd, &rect);
-	int width = rect.right;
-	int height = rect.bottom;
-
-	for (int py = 0; py < height; ++py) {
-		double y0 = ymin + (ymax - ymin) * py / height;
-		for (int px = 0; px < width; ++px) {
-			double x0 = xmin + (xmax - xmin) * px / width;
-
-			double x = 0.0;
-			double y = 0.0;
-			int iteration = 0;
-
-			while (x * x + y * y <= 4.0 && iteration < max_iter) {
-				double xtemp = x * x - y * y + x0;
-				y = 2 * x * y + y0;
-				x = xtemp;
-				++iteration;
-			}
-
-			COLORREF color = (iteration == max_iter) ? RGB(0, 0, 0) : RGB(255 - (iteration * 10) % 255, 255 - (iteration * 5) % 255, 255);
-			SetPixel(hdc, px, py, color);
-		}
-	}
-
-	EndPaint(hwnd, &ps);
-}
-
-// Function to draw the Julia set
-void DrawJulia(HWND hwnd) {
-	PAINTSTRUCT ps;
-	HDC hdc = BeginPaint(hwnd, &ps);
-
-	RECT rect;
-	GetClientRect(hwnd, &rect);
-	int width = rect.right;
-	int height = rect.bottom;
-
-	isJulia = true;
-
-	for (int py_screen = 0; py_screen < height; ++py_screen) {
-		double y0 = ymin + (ymax - ymin) * py_screen / height;
-		for (int px_screen = 0; px_screen < width; ++px_screen) {
-			double x0 = xmin + (xmax - xmin) * px_screen / width;
-
-			double x = x0;
-			double y = y0;
-			int iteration = 0;
-
-			while (x * x + y * y <= 4.0 && iteration < max_iter) {
-				double xtemp = x * x - y * y + px;
-				y = 2 * x * y + py;
-				x = xtemp;
-				++iteration;
-			}
-
-			COLORREF color = (iteration == max_iter) ? RGB(0, 0, 0) : RGB(255 - (iteration * 10) % 255, 255 - (iteration * 5) % 255, 255);
-			SetPixel(hdc, px_screen, py_screen, color);
-		}
-	}
-
-	EndPaint(hwnd, &ps);
 }
 
 // Function to draw the Mandelbrot set in a separate thread
